Add event-list to inspect finish
