+++ frontend/src/views/inspector/Inspector.vue
@@ -8,18 +8,8 @@
       <TabPane label="Real-time" name="realtime"></TabPane>
       <TabPane label="Advanced" name="advanced"></TabPane>
     </Tabs>
-    <Row v-if="selectedModeTab==='realtime'">
-      <Col :span="listSpan">
-        <flow-list class="inspector-left"></flow-list>
-      </Col>
-      <div class="split" v-if="focusedFlow"></div>
-      <Col span="12" v-if="focusedFlow">
-        <flow-detail class="inspector-right"></flow-detail>
-      </Col>
-    </Row>
-    <Row v-if="selectedModeTab==='advanced'">
-      <Events></Events>
-    </Row>
+    <FlowInspector v-if="selectedModeTab==='realtime'"></FlowInspector>
+    <EventInspector v-if="selectedModeTab==='advanced'"></EventInspector>
   </div>
 </template>
 
@@ -37,10 +27,9 @@ let recordingStatus = {
   text: 'Stop recording'
 };
 
-import FlowList from '@/views/inspector/FlowList.vue'
-import FlowDetail from '@/views/inspector/FlowDetail.vue'
 import ButtonBar from '@/views/inspector/ButtonBar.vue'
-import Events from '@/views/event/EventInspector.vue'
+import EventInspector from '@/views/event/EventInspector.vue'
+import FlowInspector from '@/views/inspector/FlowInspector.vue'
 
 export default {
   name: 'Inspector',
@@ -52,22 +41,9 @@ export default {
     };
   },
   components: {
-    FlowList,
-    FlowDetail,
     ButtonBar,
-    Events
-  },
-  computed: {
-    listSpan () {
-      if (this.focusedFlow) {
-        return '12'
-      } else {
-        return '24'
-      }
-    },
-    focusedFlow () {
-      return this.$store.state.inspector.focusedFlow
-    }
+    EventInspector,
+    FlowInspector
   },
   methods: {
     switchTab (name) {
@@ -83,12 +59,6 @@ export default {
   display: flex;
   align-items: center;
 }
-.inspector-left {
-  margin-right: 0px;
-}
-.inspector-right {
-  margin-left: 5px;
-}
 .divider {
   display: block;
   width: 100%;
@@ -96,14 +66,6 @@ export default {
   background: #eee;
   top: 0;
   left: 0;
-}
-.split {
-  display: block;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 50%;
-  border: 1px dashed #eee;
 }
 .small-tab > .ivu-tabs > .ivu-tabs-bar {
   margin-bottom: 0;
