+++ lyrebird/mock/blueprints/apis/menu.py
@@ -16,12 +16,6 @@ class Menu(Resource):
                 'path': '/'
             },
             {
-                'name': 'event',
-                'title': 'Events',
-                'type': 'router',
-                'path': '/event'
-            },
-            {
                 'name': 'datamanager',
                 'title': 'DataManager',
                 'type': 'router',
